+++ crates/re_viewer/src/ui/mod.rs
@@ -9,6 +9,7 @@ mod welcome_screen;
 pub(crate) mod add_space_view_or_container_modal;
 pub(crate) mod memory_panel;
 pub(crate) mod selection_panel;
+pub(crate) mod space_view_space_origin_ui;
 pub(crate) mod visible_history;
 
 pub use blueprint_panel::blueprint_panel_ui;
